xfs: provide a cursory analysis of next-20230725 failures
Provide a cursory analysis of the next-20230725 failures.
While at it, describe how to look for file details and cat stuff.

Later we should just add some python helpers which let's us
easily query for this, so we can search for multiple namespaces
for instance and not have to type in these long commands.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>