vagrant: make virtio drives default
virtio drives supports io-threads on qemu, this let's the storage
emulation not have to rely on same global qemu lock. We have support
and by default enable io-threads on virtio drives. Using these has
resolved issues when we use nvme on the guests instead, because nvme
on qemu lacks io-thread support. That support is coming soon but
it is not yet widespread. So for now let's use more stable stuff.

Stable XFS maintainers have been using this for a while and I just
ran some XFS tests on linux-next with it.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>