xfs: ensure low block sizes are enabled if extra storage allow
There are a few XFS profiles we allow which are section size specific,
that is the logical block size of the drive must be allow at least this
amount of block sizes. For example we cannot test a filesystem with 2k
block size on a drive which has a logical block size of 4k.

For virtualization we get to pick and choose the block sizes and this is
relevant only if you are testing with fstests and using sparsefile
generation. If you are using real NVMe drives we assume for now you
know what you are doing (ie pcie-passthrough). We also assume you know
what you are doing for now if you are using terraform.

Later on terraform users can enhance these extra storage things per
cloud provider and SKU and storage parameters.

This will prevent from allowing people bringing up nodes (virtual guests
or cloud instances) only to find out that fstests won't work for them
because the logical block size won't work. The first error message
will come up when gendsisks try to create loopback drives if using
sparsefile generation.

This also ensure that in the future if say 16k drives hit the market
(logical block size, for example NVMe format) don't support these lower
block sizes that you cannot use them there either. That applies for the
cloud and for real drives / emulation.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>