fstests: fix real nvme drive partition usage
I first added support for real nvme drives using:

   /dev/disk/by-id/nvme-eui*

symlinks and then later a fallback mechanism to look for
device model and serial number. But adding support for
the second part meant the first stopped working and
the reason is it was overloading the same register variable
and this confuses ansible. So just use separate variables.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>