fstests: propagate host specific variables on .kdevops_fstests_setup
In recent debian testing environments TEST_DEV is not being propagated
when running the ansible oscheck.sh task command. Fix this by
propagating host specific variables by using:

  parse_config_section_local <section>

Use the same heuristics to infer a section with other uses of
INFER_SECTION in other scripts.

We use /root/.kdevops_fstests_setup as a file which is sourced as root
from the root user .bashrc.

While at it, add the NFS specific variables as well, to make it
easy to add new NFS variables add a new bool for when nfs on fstests
is enabled as otherwise those variables are not defined.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>