fstests: fix running fstests-baseline when nfs server is disabled
make fstests-baseline stopped working because undefined variables
were forced to be set on the task which runs the oscheck.sh command
even if NFS was disabled. I moved setting these environment
variables on the .kdevops_fstests_setup file now, and so we can
remove these now from the environment task in ansible.

If we however determine we do need them again back then the right
way to do this is to set a default setting on the defaults.yml file
for the fstests role. But since it is not clear what that should be
just remove this as I'm pretty sure this should work now as-is with
just with new changes of using the .kdevops_fstests_setup file.

Fixes: e00143f69cdb3 ("fstests: add support for testing NFS"
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>