build_qemu: fix ignoring submodules
Oh boy, tons of changes in qemu to disable git sub modules. WTF.
Seriously, it's as if they don't give a flying fsck about users
building this.

OK hopefully this is the last thing needed to fix this.

Once fedora and SUSE tumbleweed picks up a package for b4 we
can just use that too there.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>