nfsd: add a second export
Running fstests on nfs vs. a single export is problematic, as the test
and scratch mounts can end up sharing a superblock. Using the same
underlying fs for 2 mounts may make it difficult to test fscache in the
future as well.

Claim a second extra_disk on the NFS server to act as a second exported
filesystem. Rejigger the pynfs and fstests workflows to take advantage
of the new export.

While we're there, rework the Kconfig defaults to use by-id device names
on vagrant. For now, leave the ultimate default as a blank string, since
terraform configs can vary widely.

Note that this may be problematic if you try to run a dedicated pynfs
workflow on terraform, since a lot of those terraform configs only
provision 2 extra disks per guest, and this requires 3.

Signed-off-by: Jeff Layton <jlayton@kernel.org>