vagrant: rework Fedora kconfig options
The Kconfig options for Fedora releases are a bit too complex. We don't
really need a radio button for this. Just have it default to the latest
release string if VAGRANT_FEDORA is set, and then add a new boolean
option that defaults to off for old versions that handles the FSTESTS_XFS
selects properly.

Signed-off-by: Jeff Layton <jlayton@kernel.org>