bootlinux: add support for "config-kdevops"
Currently, you're expected to have a kernel config file with a name
like "config-{{ target_linux_config }}". If there isn't one, then the
play aborts.

For my use-case, I just want to have a stock config that I periodically
update and use regardless of the branch name. If there is a file called
config-kdevops in the template directory, use that instead of any
version-specific branch.

Reviewed-by: Luis Chamberlain <mcgrof@kernel.org>
Signed-off-by: Jeff Layton <jlayton@kernel.org>