vagrant: allow the aio= and cache= options to be set on any drive type
The aio= and cache= options to the -drive option don't have any
dependency on the interface that is presented to the guest.

Rename *_VIRTIO_AIO_* to *_AIO_*, and remove the dependency on virtio in
the Kconfig. Add the appropriate cache= and aio= options in the
Vagrantfile template.

While we're in here, let's add support for io_uring on an experimental
basis.

Reviewed-by: Luis Chamberlain <mcgrof@kernel.org>
Signed-off-by: Jeff Layton <jlayton@kernel.org>