fstests: remove redundant string prompts for fstype
We use a radio button choice selector to decide what filesystem to use
for these options, so we don't want to present a string prompt. Doing so
means that we have to change the configuration in two places if you want
to switch to testing a different fileystem.

Remove the prompt strings to hide these options altogether.

Signed-off-by: Jeff Layton <jlayton@kernel.org>