vagrant: remove redundant string for cpu and mem choice
This commits is part of the propagation fix found by Jeff
Layton in the choice/endchoice Kconfig option. When a CONFIG option
is used to assign the selected CONFIG, a string can't be added to
that config, otherwise the config is exposed and not updated with
user selection.

Remove the string part of the int CONFIG_ in mem and cpu options so it
gets hidden from the menuconfig.

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>