fstests: Request inferred user
Currently when running make fstests for bare-metal setups, the data_user
'vagrant' is set for the /data partition. This is currently set as the default
data_user. We have the option to edit this in extra_vars.yml.

make linux checks for an inferred user prior to setting the uid/gid for
the /data partition.

In the situation where we don't add the 'vagrant' user to the
bare-metal, we get a "user does not exist" on make fstests.

In both cases we do not want to switch users between the two targets,
nor do we want to necessarily ensure that the 'vagrant' user exists on
the metal. Therefore, request the inferred user on make fstests.

Signed-off-by: Frederick Lawler <fred@cloudflare.com>
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>