nfsd: new scheme for NFS exports
The current scheme for NFS exports on the server is problematic. We
currently only 2 filesystems, and the clients create and use
subdirectories under that.

The problem there is that some fstests like to fill up the scratch
partition. If we're running fstests in parallel, this can cause tests
on another host to fail. It can also take quite some time to fill up
100G.

Drop the filesystems on the raw block devices and convert to LVM2. Add a
new playbook that will create an LVM volume on the nfs server, put a
filesystem on it, mount it and export it.

Rework the pynfs and fstests workflows to create new, dedicated exports
for their testing.

Signed-off-by: Jeff Layton <jlayton@kernel.org>