fstests: add support for START_AFTER
To help with new baselines initial support was added a while ago
to kdevops to use the check.time by using

make fstests-baseline INITIAL_BASELINE=1

file but turns out that relying on the check.time file is stupid
as it can often be empty. We don't populate it as we go at all
so this is useless for an initial baseline.

So instead extend the arguments added for this to just add a new
simple ./check --start-after <test-which-fails>. And document all
this.

Fixes: 8cbd3314da0415f7 ("fstests: add support to use check.time files as expunges"
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>