xfs: fix regressions and few crashes
After leaving a few system tests running overnight a fix now
merged fixes the only 3 regressions for xfs_reflink_4k. It so
happens that this also fixes 4/6 of the crashes we had. Those
test which used to crash were left running overnight on a system
in a loop so it ran about 40 or 50 times.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>