terraform: gce: remove redundant string for cpu and mem choice
This commits is part of the propagation fix found by Jeff
Layton in the choice/endchoice Kconfig option. When a CONFIG option
is used to assign the selected CONFIG, a string can't be added to
that config, otherwise the config is exposed and not updated with
user selection. Removing the string makes the config option hide from
the menuconfig.

For multichoice menu with single choice, remove choice menu and
leave it as standalone option with its default value.

Also, s/gce/GCE/g for Google Compute Engine (GCE) or Google Cloud
Platform Provider (GCP) and add documentation link to the terraform
variable each option is assigned to.

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>