xfs: add 32k and 64k LBS initial expunges
In order to test 32k LBS and 64k LBS we need an initial baseline.
For now just use the 16k baseline as a symlink. If we see too
many differences we will have to create a new file for each section.

Note that XFS does not support block size 64k with 64k sector size.
Since I will be testing a real NVMe drive with 64k I'll use a 4k
sector size for now to test 64k. We need a new profile for 64k_32ks.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>