xfs: extend next-20230825 for rtdev
These new failures happen only with a low failure rate of about
1 out of 70 runs or so. Upon inspection of the failure, it seems
likely a test bug, this requires some manual loop testing on rtdev
to see why exactly this is failing. But the lack of any dmesg
entries seems to indicate nothing kernel related here, likely
a test bug.

Proactively also add it to rtlogdev section as well as this may
be rt related.

https://gist.github.com/mcgrof/125e451cedeb584c0513d454d8ba2e30

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>