xfs: add expunge list for 32k and 64k_4ks
It turns out that there are *real test bugs* for only >= 32k sector
sizes. Most failures are test bugs!

There are however also some  *real-bugs* for 32k and also for
64k block sizes as well.

Perhaps of more interest to us should be xfs/166 as it happens on both
and it seems real.

Worth noting is how 64k is being tested with 4k sector size, the
64k_32ks profiles will need to inherit the 32k test bug baselines
as well.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>