linux: speed up regular commands
Commands such as:

make linux-deploy
make linux-install
make linux-install
make linux-reboot
make uname

are supposed to be *fast*, and having them depend on the stupid
nodes file makes it slow. Yes we need the file, but you should already
have it if you ran 'make linux'. So accept the huge speed up boost over
the stupid carefulness we are using to verify the files are present for
the nodes.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>