workflows: make kdevops custom repo optional and add version support
When user modifies WORKFLOW_KDEVOPS_GIT, it is no longer autogenerated
as it's an exposed config. Making it optional and hiding behind a new
custom option guarantees a proper default value when user disables
the custom option.

Add optional custom kdevops repository option
(WORKFLOW_KDEVOPS_GIT_CUSTOM) to let the user decide which repo is
cloned on the targets. In addition, if the new custom is not selected,
then WORKFLOW_KDEVOPS_GIT option is non-visible, so it always generates
the desired value.

Add WORKFLOW_KDEVOPS_GIT_VERSION to control the version of the
repository to check out on the target. By default, ansible uses HEAD if
not provided, so keep it as is by default and let the user redefine it.

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>