fstests: make device pool strategy conditional
In preparation for running tmpfs fstests, make the device pool strategy
conditional. If the strategy is not defined, do not run ansible scripts
that requires a strategy. For that, is not enough anymore to check for
the strategy selected but how the other options are also configured
making them explicit in the when condition.

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>