tmpfs: add nameless (no description) tmpfs bool enable
It is easier to read ansible with bools and these are best then
described with what they enable. Today we do manual Makefile
hacking to add bools to ansible, but this is silly to scale.
To scale we will simply later add support to modify Kconfig
so we can opt-in which Kconfig symbols could be also output
on an extra_vars.yaml file. This would then let the extra_vars.yaml
file be a small as it is today, contrary to our large .config.
It also means, more importantly, that in the future all the
Makefile junk to check for config stuff could be removed as it
would be done automatically for us. So to make that long term
goal easier, we can just try to match kconfig symbols to ansible
bools. That means adding a nameless bool for tmpfs then so
CONFIG_FSTESTS_TMPFS_ENABLE ands up in an equivlent automatic
fstests_tmpfs_enable.

Spell this out a bit so to help others help out with scaling
in the future, and understand why we add these, what seem to
be, superfluous extra bools on Kconfig. Adding an extra Kconfig
bool is easy just make it clearer what the goals are with ansible
variables.

While at it, remove the mention of how tmpfs used to be called
shmem as it can create confusion. tmpfs is what we expose to
userspace, if you don't enable the user support to use shmem
then you get only shmem support in the kernel. Ie, tmpfs extends
shmem support to allow users to go willy nilly with it as a
filesystems with options.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>