scripts/vagrant.Makefile: skip post bringup after running it once
For vagrant environments running 'make bringup' does
$(cd vagrant && vagrant up) and then we kick optionally kick
off ansible to update our ssh configuration and then run some
post provisioning tasks. The ssh config update and post
provisioning only needs to be run once so put up a guard
that ensure this so we can just run 'make bringup' after
for example we reboot a host system.

This allows us to be consistent for bringups and let's folks
not familiar with vagrant to just use that target.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>