xfs: augment expunges for v6.6.0-rc5 with a new failures
xfs/438 can end up in a hung task on two sections on v6.6-rc5 with
a failure rate of about ~ 1/2, in other words you need to run the
test at least twice. The failed sections are:

  * xfs_crc_logdev
  * xfs_reflink_logdev

We provide gists to the hung tasks, each of which are collected directly
from the respective guests on the respective section.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>