xfs: augment 6.6.0-rc5 baseline with low failure rate non-fatal failures
There are a few tests which require running at least about 3 times
before seeing a failure, expunge them and document their approximate
failure rate.

In the future, what we want, is to come up with a syntax for these so
we can verify if the failure rate remains the same or gets worse, or
if they no longer fail. So for now just add a new tag to query for
these easily. For now these tags are being added manually but later
in the future we could just use the .kernelci* files to gather the
loop number for these.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>