xfs: add archive for second cluster run on v6.6-rc5
There is a second cluster running XFS tests on v6.6-rc5 this
collects the results there. This the 7th manual loop I run, and
so failure rate is about 1/7, the prior commit documents those
failures and respective expunges. Folks wanting to look at the actual
results can inspect the tarball and cluster configuration for kdevops
here.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>