xfs: use v6.6-rc5 expunges as base for 6.6.0-rc5-large-block-minorder
We want to first identify regressions, so use the v6.6-rc5 baseline for
6.6.0-rc5-large-block-minorder. We can later also re-test all known
failures on the baseline as well, to see and verify if non-fatal tests
are now fatal / cause a crash but we can do that as secondary step.

It is much easier to detect regressions with just:

   make fstests-baseline

Then to start from scratch, and re-test known failures later. The
confidence in the baseline for v6.6-rc5 is at loop 10 at this point,
so this means if we find new failures on that directory expunge we
will need to also expunge it now on the minorder expunge directory too.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>