xfs: add first large-block-minorder fatal failures
Upon testing large-block-minorder branch of Linux on kdevops [0] with
top commit id 67dc16aea31 (" kdevops: add localversion-dev tag for
kdevops and changelog") this is the first set of fatal failures found
with only 3 new failures:

  1) xfs_reflink_4k: xfs/168 hung task
  2) xfs_nocrc_32k_4ks: xfs/041 xfs assert fs/xfs/xfs_message.c:102
  3) xfs_reflink_64_4ks: xfs/599 hint suggests that xfsprogs
     commit b7b81f336ac ("xfs_repair: fix incorrect dabtree hashval
     comparison" might fix this.

Failure 1) seems to be a recurring issue and likely we just have not
triggered it without minorder on xfs_reflink_4k.

Failure 2) is a common failure already observed on other test sections so
also likely something we could possibly trigger without min order.

The only new failure I can see at the moment is 3) xfs/509 but we
should re-test and see if that is also possibly triggered on 64k
PAGE_SIZE system. The xfs/509.hints suggests this is also already
fixed on xfsprogs, so we should verify that.

[0] https://github.com/linux-kdevops/linux/tree/large-block-minorder
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>