xfs: add first series of crashes with SOAK_DURATION=9900 on v6.6-rc5
Enabling SOAK_DURATION=9900 proves to crash *most* sections, all
except 3 sections crashed on generic/475 due to the pattern:

xfs_inodegc_worker() -->
	xfs_ifree() -->
		--> XFS: Assertion failed: ip->i_nblocks == 0, file: fs/xfs/xfs_inode.c, line: 2359

And so then kernel BUG at fs/xfs/xfs_message.c:102.

The test section xfs_reflink_2k was the only test section that managed
to pass this test, but for all intents and purposes it likely could fail
if we ran it longer.

Although we are using the all.txt file, by no means is this a
lazy-baseline, this is a real valid use case for the all.txt file
which reflects a common issue present on XFS today on v6.6-rc5. Just
20 minutes with generic/475 should do the trick.

The second crash happens on xfs_crc_rtdev_extsize_28k on generic/388
on a slightly similar pattern starting on xfs_inodegc_worker() but
now with xfs_iext_set() instead of xfs_ifree() and with a different
assert.

xfs_inodegc_worker() -->
	xfs_iext_set() -->
		--> XFS: Assertion failed: (irec->br_blockcount & ~XFS_IEXT_LENGTH_MASK) == 0, file: fs/xfs/libxfs/xfs_iext_tree.c, line: 58

The third crash happens on xfs_reflink_2k, which managed to pass
generic/475, but now fails on generic/648 on the exact same pattern
as in the generic/475 failures ending in the assert for
ip->i_nblocks == 0, on fs/xfs/xfs_inode.c, line: 2359.

My guess is that the generic/648 crash will also be triggered
on the other test sections as well and we'll just have to move this
test failure onto all.txt on v6.6-rc5 and the min order baseline.
The only reason we don't yet have it on all.txt is because obviously
the other guests crashed first on generic/475.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>