fstests_watchdog.py: enhance kernel ring buffer scraper
fstests_watchdog.py relies on dmesg output, however the default
kernel ring buffer is of size LOG_BUF_SHIFT is set to 17 so
that is only 128 KiB. The kernel used on kdevops for v6.6-rc5
is also setting CONFIG_LOG_BUF_SHIFT=17 so again only 128 KiB.

This is too damn small when testing with large SOAK_DURATION values
such as 9900 (2.5 hours) and what will happen is the ring buffer
loops around so that the head no longer display any fstests preamble
messages such as "run fstests". For instance:

vagrant@line-xfs-reflink-4k ~ $ sudo head /dev/kmsg | head -1
6,123134,53983829503,-;loop0: detected capacity change from 0 to 13648394

The fstests_watchdog.py relies on dmesg to scrape for the last
fstests test run, however the above limitation means it will think
no fstests test has ever run on this system. To fix this while also
remaining compatible with non-systemd systems, we look for journalctl
and if present rely on that for both looking and also grep'ing of
"run fstests".

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>