fstests_watchdog.py: use SOAK_DURATION
The watchdog relies on a series of heuristics to determine if a
test has crashed. If you already ran a test the prior run time
test is used. If no prior runtime for a test is known then we
use a kdevops configuration variable. However we don't use the
SOAK_DURATION today.

Although, as the prior commit clarifies, SOAK_DURATION is only
used for certain tests, and so it is silly to use expand all test
coverage to assume any test can take that long. Instead just keep
a current static list which we can expand on as we see fit. In the
future we can build this list dynamically based on the fstests git
tree, or have the fstests git tree have a list of known tests which
use it.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>