xfs: xfs/599 on LBS
xfs/599 tests clashes on hashes used for metadata dump, however its
efforts to create xattrs based on the blocksize proves to increase
the amount of time it takes to complete exponentially, on 32k block
size its about 30 hours.

For details see:

https://lore.kernel.org/fstests/ZULu%2FRm%2FEiBY8ZzG@bombadil.infradead.org/T/#u

For now expunge this, even though the test actually passes. It just
takes too damn long and the solution is not clear yet.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>