terraform: make data device provider specific
Move the data device definition to each provider's Kconfig file just
as with OCI. This also sets the /dev/nvme1n1 to be the default data
device for AWS, before this nothing was set.

This also fixes the default setting of the data device using the
automatic kconfig magic when no prompt is set on a string, just as
we have been fixing other strings in kdevops before.

We can now use c7a.metal-48xl for example with: make linux

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>