guestfs: add a new local guest management variant
The future of Vagrant is somewhat uncertain, so we need to replace it.

This patch adds a new way to manage local guests, using libguestfs-tools:

    https://libguestfs.org/

At a high level, this new option mostly affects the "make" and "make
bringup" stages. Instead of using vagrant, it will:

- use gen_nodes to build a nodelist and a libvirt xml file for each
  guest (instead of building a Vagrantfile)
- build a base image using virt-builder for a given distro (see
  virt-builder -l for the list of options)
- clone that image (using reflink) to a base root image for each guest
- create extra disks for each guest
- define the new guests using the xml file and start them

While no one really loves XML, I think it turns out to be a lot simpler
than trying to template out a ruby script (like we do with the
Vagrantfile.

Not all storage and hardware configurations are yet supported. I'm
hoping others that are using those configurations will help flesh that
out in later patches.

Signed-off-by: Jeff Layton <jlayton@kernel.org>