devconfig: fix when: checks about custom Fedora repofile
Without this change, I get:

fatal: [kdevops-nfsd -> localhost]: FAILED! => {"changed": false, "msg": "argument 'path' is of type <class 'NoneType'> and we were unable to convert to path: 'None' is not a string and conversion is not allowed"}

Apparently having this value set to "None" makes it "defined". Change
it to just check the bare value and evaluate that to true or false.

Signed-off-by: Jeff Layton <jlayton@kernel.org>