kconfig: move libvirt settings to a new file
The the libvirt Kconfig options that have VAGRANT in the name are really
all about libvirt and have little to do with vagrant at all. While we
need to replace vagrant, we want to keep a lot of the existing libvirt
handling code intact. This patch renames most of the options that have
VAGRANT_LIBVIRT into the name to just be LIBVIRT.

Move all of the libvirt kconfig settings to a new file, and rename the
ones that have VAGRANT in the name.

There should be no functional changes, but the Kconfig symbol renaming
is likely to break existing Kconfigs. If you're using Vagrant, be sure
to regenerate your config file after this change.

Signed-off-by: Jeff Layton <jlayton@kernel.org>