workflows/cxl: add gen_qemu_cxl.py
This script generates QEMU parameter for a user-specified CXL
configuration, in either cmdline or xml format.

The script is not yet generated into the cxl workflow, but that
will be a next step. For now we put it in kdevops to encourage
testing and feedback (even outside of kdevops itself).

TODOs:
* Fix permissions for created memdev files to match what will be
  needed for libvirt (right now it just sets all files to 0666).
* Support memdevs connected directly to CXL root port - currently
  it only supports a switch connected to the root port.
* Remove --bus-alloc-per-host-bridge parameter, and just figure
  out number of buses needed based on user-specified parameters.

Signed-off-by: Jim Harris <jim.harris@samsung.com>