kdevops: add support for remote journals
It is a royal pain to debug nodes either hosted virtually or on the
cloud when doing kernel testing. Consoles have helped but in order to
scale and increase automation we need to instead start dumping our
journals over the wire on the fly because many times a crash can
leave a host really inaccessible even with libvirt virsh console.

Since most modern distros use systemd, leverage systemd journal for
this exact purpose as it already has a built-in mechanism for this
[0] [1] [2].

So we add a few new commands:

make journal-server      - Setup systemd-journal-remote on localhost
make journal-client      - Setup systemd-journal-remote on clients
make journal-ls          - List journals available and sizes
make journal-ln          - Add symlinks with hostnames

The last one is needed because even if we install libvirt-nss [4] on
the host and use it on nsswitch.conf we still cannot get the host
to do reverse lookups. It seems one option may be modifying the
network XML but for that we can look towards libguest instead to
correct this on new bring ups.

If you are struggling with getting crashes from nodes and you already
had a big cluster running, don't be afraid, just git pull and enable
CONFIG_DEVCONFIG_ENABLE_SYSTEMD_JOURNAL_REMOTE and ensure that
CONFIG_DEVCONFIG_SYSTEMD_JOURNAL_REMOTE_URL is correct, and then run:

make
make journal-server journal-client

Give it a few, and after a while:

make journal-ls

Once you see the count matches your nodes, use the final hack to
get real hostname symlinks for them:

make journal-ln

And hack away now!

[0] https://www.freedesktop.org/software/systemd/man/latest/systemd-journal-remote.service.html
[1] https://www.freedesktop.org/software/systemd/man/latest/systemd-journal-upload.service.html
[2] https://www.freedesktop.org/software/systemd/man/latest/journal-upload.conf.html
[4] https://libvirt.org/nss.html

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>