nfsd_add_export: Enable nfsd_add_export to set export ownership
The two workflows that currently use kdevops-nfsd run as root, so
the ownership of kdevops-nfsd hasn't mattered so far. But I'm about
to add a new workflow that runs as a regular user. kdevops-nfsd
exports will need to be owned by the user ID that runs that workflow.

There should be no change for the fstests and pynfs workflows: the
default owner of export_volname will continue to be "root:root".

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>