kdevops: add journal-restart and journal-status
While testing systemd-remote-journal I thought that it was not working
in realtime, it was not correct, the issue was that the upload service
had not started correctly on the guest. So add to new targets which
lets us restart and also check the status of the service. The restart
is only done for the clients, as the server is on the localhost and
you can use systemd manually for that.

  make journal-restart - restarts upload service on client
  make journal-status  - checks status on server and client

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>