systemd_journal_remote: add sane permissions
Right now to scape a journal we call sudo each time which is just
an abuse and also given we run a watchdog will make the privileged
logs lengthy. To fix this we must use sensible permissions for the
directory, and so we use what systemd uses by default and add the
tasks to ensure the user is part of the systemd-journal-remote group.

If you already have this set up just run:

make journal-server

While at it, I see there may be some broken symlinks so remove them.
And the make journal-ls is a bit too chatty, so shorten it to what
we want.

While at it, enhance scripts/workflows/fstests/fstests_watchdog.py so
that it verifies it can read without sudo and otherwise complain. So
then remove the sudo prefix for its read of the journal.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>