Enable NFS/UDP testing
NFSv3 over UDP is still used in some environments. Enable it in
the local NFS server configuration so that target hosts may mount
via UDP.

Note that the Linux kernel Kconfig option

  CONFIG_NFS_DISABLE_UDP_SUPPORT

must be set to N (or left unset) in order for UDP to be built into
the in-kernel NFS client on the target hosts.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>