bringup: Add soft iWARP bring-up option
An iWARP device on the target hosts enables testing RDMA-based
storage protocols such as SMB Direct, NVMeoF, or NFS/RDMA.

"rdma link" has to be done after every boot; it's not sticky, since
there's no physical RDMA device involved. Thus a udev rule is
inserted on each host to instantiate the siw device for the local
eth0, whenever it is rebooted.

Note that the Linux kernel Kconfig option

  CONFIG_RDMA_SIW

must be set to M or Y in order for the siw.ko driver to be built for
the target hosts.

Suggested-by: Leon Romanovsky <leon@kernel.org>
Tested-by: Jeff Layton <jlayton@kernel.org>
Reviewed-by: Jeff Layton <jlayton@kernel.org>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>