mirror: add a smart git check
If you are on a firewall chances are you will fail miserably with the
default mirroring options and this is a terrible experience as the
git protocol will fail. The git protocol is nice and efficient so
we don't want to change the default to https.

So add a smart git check on make menuconfig which will test if the git
protocol will work, otherwise disable all git options. This incurs
a penalty on checking if the git protocol works once on make menuconfig
with a timeout of 3 seconds. This hack only works on systems which have
netcat installed, otherwise we assume git will work too. Ie, if you want
this smart heuristic install netcat.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>