bringup: split provisioning into 2 steps
Right now we deal with provisioning as a post bringup work item and it
does two things:

1) Update ssh config file
2) Runs the devconfig playbook, which optionally installs some deps
   other than do basic things like setup console, kdump, etc.

The fact that deps are optional on the devconfig playbook makes it
hard to add dependencies we want to install without incurring a penalty
to everyone as a default option to install all standard generic kernel
development tools we define under CONFIG_KDEVOPS_TRY_INSTALL_KDEV_TOOLS.
The reason we disable CONFIG_KDEVOPS_TRY_INSTALL_KDEV_TOOLS by default
is it delays bringup.

However the devconfig playbook is also used to place random oddball
generic tools may want to install on bringup. Some of these tools we
may want to try to install early on, even before we run the devconfig
playbook for generic stuff like setting up the console or kdump.

To help with this, and in order to allow us to keep relying on the
devconfig playbook for random packages we may want to allow users
to install, we split the above two provisioning tasks into two
separate tasks. This will allow us to add another middle set of
targets which get executed after 1) but before 2). We will add
support for this in the next subsequent commit.

For now this should be a no-op for both vagrant and libguest. For
terraform we actually enhance the experience by adding a guard to
not run the devconfig playbook twice if a user by mistake tried to
run 'make bringup' twice. This guard was not present before for
terraform.

Maybe later on we should share these snippets between all 3 bringup
methods. Note we have no generic task for bare metal yet, but we
should -- we should just add the devconfig playbook. That would
cases where bare metal systems which are already brought up and
you have ssh configuration going for already working.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>